ods_airpass/requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.2
django-admin-interface==0.29.1
django-colorfield==0.11.0
django-import-export==4.2.0
mysqlclient==2.2.5
pillow==11.0.0
python-slugify==8.0.4
sqlparse==0.5.1
tablib==3.7.0
text-unidecode==1.3
tzdata==2024.2
